CASC-77
removed the CleanUpRegistry, Cleanable interface, etc. and just use a simple static getter on the Filter that returns the Storage. The TimerTask just calls this static getter directly. This was deemed simpler than the CleanUpRegistry solution.

for Spring config: rather than use Quartz, we just use the java.util.Timer support in Spring