Implemented URL Exclusion Pattern capabilities for AuthenticationFilter class
I have found that it is useful for the AuthenticationFilter to allow for setting up of certain
URL patterns to be excluded from filtering and have such setup the filter to allow for
the user to specify, using init-params, a comma separated list of regex values which will
indicate what url patterns should not be sent through the filter. This was meant to solve a
problem that existed when used in gateway mode and AJAX calls were attempted through
the filter, if the user was not logged in the CAS redirection would not work properly and the
AJAX call would never receive to return value that was expected.

In addition I updated the Test cases for the AuthenticationFilter to utilize JUnit 4 annotations
as well as adding mockito to the dependencies for testing to allow for easily mocking classes.